+++ skillfarm/app_settings.py
@@ -2,13 +2,8 @@
 App Settings
 """
 
-# Standard Library
-import sys
-
-# Alliance Auth (External Libs)
-from app_utils.app_settings import clean_setting
-
-IS_TESTING = sys.argv[1:2] == ["test"]
+# Django
+from django.conf import settings
 
 # EVE Online Swagger
 EVE_BASE_URL = "https://esi.evetech.net/"
@@ -29,27 +24,30 @@ ZKILLBOARD_KILLMAIL_URL_REGEX = r"^http[s]?:\/\/zkillboard\.com\/kill\/\d+\/"
 # Set Test Mode True or False
 
 # Set Naming on Auth Hook
-SKILLFARM_APP_NAME = clean_setting("SKILLFARM_APP_NAME", "Skillfarm")
+SKILLFARM_APP_NAME = getattr(settings, "SKILLFARM_APP_NAME", "Skillfarm")
 
 # If True you need to set up the Logger
-SKILLFARM_LOGGER_USE = clean_setting("SKILLFARM_LOGGER_USE", False)
+SKILLFARM_LOGGER_USE = getattr(settings, "SKILLFARM_LOGGER_USE", False)
 
 # Max Time to set Char Inactive
-SKILLFARM_CHAR_MAX_INACTIVE_DAYS = clean_setting("SKILLFARM_CHAR_MAX_INACTIVE_DAYS", 3)
+SKILLFARM_CHAR_MAX_INACTIVE_DAYS = getattr(
+    settings, "SKILLFARM_CHAR_MAX_INACTIVE_DAYS", 3
+)
 
 # Batch Size for Bulk Methods
-SKILLFARM_BULK_METHODS_BATCH_SIZE = clean_setting(
-    "SKILLFARM_BULK_METHODS_BATCH_SIZE", 500
+SKILLFARM_BULK_METHODS_BATCH_SIZE = getattr(
+    settings, "SKILLFARM_BULK_METHODS_BATCH_SIZE", 500
 )
 
 # Set Notification Cooldown in Days
-SKILLFARM_NOTIFICATION_COOLDOWN = clean_setting("SKILLFARM_NOTIFICATION_COOLDOWN", 3)
-
+SKILLFARM_NOTIFICATION_COOLDOWN = getattr(
+    settings, "SKILLFARM_NOTIFICATION_COOLDOWN", 3
+)
 # Global timeout for tasks in seconds to reduce task accumulation during outages.
-SKILLFARM_TASKS_TIME_LIMIT = clean_setting("SKILLFARM_TASKS_TIME_LIMIT", 7200)
+SKILLFARM_TASKS_TIME_LIMIT = getattr(settings, "SKILLFARM_TASKS_TIME_LIMIT", 7200)
 
 # Price Source default is Jita
-SKILLFARM_PRICE_SOURCE_ID = clean_setting("SKILLFARM_PRICE_SOURCE_ID", 60003760)
+SKILLFARM_PRICE_SOURCE_ID = getattr(settings, "SKILLFARM_PRICE_SOURCE_ID", 60003760)
 
 SKILLFARM_STALE_TYPES = {
     "skills": 30,